Sejik/web_test/backend/requirements.txt
aiosmtplib==3.0.2
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.7.14
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.116.1
fastapi-mail==1.5.0
fastapi-sso==0.18.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
oauthlib==3.3.1
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
python-multipart==0.0.18
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.42
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1

# IBM Watson Machine Learning SDK
ibm-watson-machine-learning==1.0.368

# OCR Dependencies
pytesseract==0.3.13
Pillow==10.4.0
PyPDF2==3.0.1

